+++ gnubg/render.h
@@ -1,11 +1,11 @@
 /*
- * render.h
+ * Copyright (C) 1997-2002 Gary Wong <gtw@gnu.org>
+ * Copyright (C) 2003-2015 the AUTHORS
  *
- * by Gary Wong <gtw@gnu.org>, 1997-2002.
- *
- * This program is free software; you can redistribute it and/or modify
- * it under the terms of version 3 or later of the GNU General Public License as
- * published by the Free Software Foundation.
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
@@ -13,8 +13,7 @@
  * GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  *
  * $Id$
  */
@@ -82,8 +81,8 @@ typedef struct _renderdata {
     lighttype lightType;        /* Positional/Directional light source */
     float lightPos[3];          /* x,y,z pos of light source */
     int lightLevels[3];         /* amibient/diffuse/specular light levels */
-    int boardAngle;             /* Angle board is tilted at */
-    int skewFactor;             /* FOV adjustment */
+    float boardAngle;           /* Angle board is tilted at */
+    float skewFactor;           /* FOV adjustment */
     int planView;               /* Ortho view? */
     float diceSize;             /* How big are the dice */
     int roundedEdges;           /* Rounded board edges? */
